# Add fast and medium paths for decoding of floats and doubles