--- a/zio-json/shared/src/test/scala/zio/json/internal/SafeNumbersSpec.scala
+++ b/zio-json/shared/src/test/scala/zio/json/internal/SafeNumbersSpec.scala
@@ -8,88 +8,99 @@ import zio.test._
 object SafeNumbersSpec extends ZIOSpecDefault {
   val spec =
     suite("SafeNumbers")(
-      test("valid big decimals") {
-        check(genBigDecimal)(i => assert(SafeNumbers.bigDecimal(i.toString, 2048))(isSome(equalTo(i))))
-      },
-      test("invalid big decimals") {
-        val invalidBigDecimalEdgeCases = List(
-          "N",
-          "Inf",
-          "-NaN",
-          "+NaN",
-          "e1",
-          "1.1.1",
-          "1 ",
-          "NaN",
-          "Infinity",
-          "+Infinity",
-          "-Infinity"
-        ).map(s => SafeNumbers.bigDecimal(s))
+      suite("BigDecimal")(
+        test("valid big decimals") {
+          check(genBigDecimal)(i => assert(SafeNumbers.bigDecimal(i.toString, 2048))(isSome(equalTo(i))))
+        },
+        test("invalid big decimals") {
+          val invalidBigDecimalEdgeCases = List(
+            "N",
+            "Inf",
+            "-NaN",
+            "+NaN",
+            "e1",
+            "1.1.1",
+            "1 ",
+            "NaN",
+            "Infinity",
+            "+Infinity",
+            "-Infinity",
+            "1eO",
+            "1e+2147483648",
+            "1e+3147483648",
+            "9" * 99,
+            "0." + "9" * 99
+          ).map(s => SafeNumbers.bigDecimal(s))
 
-        assert(invalidBigDecimalEdgeCases)(forall(isNone))
-      },
-      test("valid big decimal edge cases") {
-        val invalidBigDecimalEdgeCases = List(
-          ".0",
-          "-.0",
-          "0",
-          "0.0",
-          "-0.0", // zeroes
-          "0000.1",
-          "0.00001",
-          "000.00001000" // various trailing zeros, should be preserved
-        )
-
-        check(Gen.fromIterable(invalidBigDecimalEdgeCases)) { s =>
-          assert(SafeNumbers.bigDecimal(s).get.compareTo(new java.math.BigDecimal(s)))(equalTo(0))
-        }
-      },
-      test("invalid BigDecimal text") {
-        check(genAlphaLowerString)(s => assert(SafeNumbers.bigDecimal(s))(isNone))
-      },
-      test("valid BigInteger edge cases") {
-        val inputs = List(
-          "00",
-          "01",
-          "0000001",
-          "-9223372036854775807",
-          "9223372036854775806",
-          "-9223372036854775809",
-          "9223372036854775808"
-        )
-
-        check(Gen.fromIterable(inputs)) { s =>
-          assert(SafeNumbers.bigInteger(s))(
-            isSome(
-              equalTo(new java.math.BigInteger(s))
-            )
+          assert(invalidBigDecimalEdgeCases)(forall(isNone))
+        },
+        test("valid big decimal edge cases") {
+          val invalidBigDecimalEdgeCases = List(
+            ".0",
+            "-.0",
+            "0",
+            "0.0",
+            "-0.0", // zeroes
+            "0000.1",
+            "0.00001",
+            "000.00001000" // various trailing zeros, should be preserved
           )
-        }
-      },
-      test("invalid BigInteger edge cases") {
-        val inputs = List("0foo", "01foo", "0.1", "", "1 ")
 
-        check(Gen.fromIterable(inputs))(s => assert(SafeNumbers.bigInteger(s))(isNone))
-      },
-      test("valid big Integer") {
-        check(genBigInteger)(i => assert(SafeNumbers.bigInteger(i.toString, 2048))(isSome(equalTo(i))))
-      },
-      test("invalid BigInteger") {
-        check(genAlphaLowerString)(s => assert(SafeNumbers.bigInteger(s))(isNone))
-      },
-      test("valid Byte") {
-        check(Gen.byte(Byte.MinValue, Byte.MaxValue)) { b =>
-          assert(SafeNumbers.byte(b.toString))(equalTo(ByteSome(b)))
+          check(Gen.fromIterable(invalidBigDecimalEdgeCases)) { s =>
+            assert(SafeNumbers.bigDecimal(s).get.compareTo(new java.math.BigDecimal(s)))(equalTo(0))
+          }
+        },
+        test("invalid BigDecimal text") {
+          check(genAlphaLowerString)(s => assert(SafeNumbers.bigDecimal(s))(isNone))
         }
-      },
-      test("invalid Byte (numbers)") {
-        check(Gen.long.filter(i => i < Byte.MinValue || i > Byte.MaxValue)) { b =>
-          assert(SafeNumbers.byte(b.toString))(equalTo(ByteNone))
+      ),
+      suite("BigInteger")(
+        test("valid BigInteger edge cases") {
+          val inputs = List(
+            "0",
+            "0123",
+            "-123",
+            "-9223372036854775807",
+            "9223372036854775806",
+            "-9223372036854775809",
+            "9223372036854775808"
+          )
+
+          check(Gen.fromIterable(inputs)) { s =>
+            assert(SafeNumbers.bigInteger(s))(
+              isSome(
+                equalTo(new java.math.BigInteger(s))
+              )
+            )
+          }
+        },
+        test("invalid BigInteger edge cases") {
+          val inputs = List("0e+1", "01E-1", "0.1", "", "1 ")
+
+          check(Gen.fromIterable(inputs))(s => assert(SafeNumbers.bigInteger(s))(isNone))
+        },
+        test("valid big Integer") {
+          check(genBigInteger)(i => assert(SafeNumbers.bigInteger(i.toString, 2048))(isSome(equalTo(i))))
+        },
+        test("invalid BigInteger") {
+          check(genAlphaLowerString)(s => assert(SafeNumbers.bigInteger(s))(isNone))
         }
-      },
-      test("invalid Byte (text)") {
-        check(genAlphaLowerString)(b => assert(SafeNumbers.byte(b.toString))(equalTo(ByteNone)))
-      },
+      ),
+      suite("Byte")(
+        test("valid Byte") {
+          check(Gen.byte(Byte.MinValue, Byte.MaxValue)) { b =>
+            assert(SafeNumbers.byte(b.toString))(equalTo(ByteSome(b)))
+          }
+        },
+        test("invalid Byte (numbers)") {
+          check(Gen.int.filter(i => i < Byte.MinValue || i > Byte.MaxValue)) { b =>
+            assert(SafeNumbers.byte(b.toString))(equalTo(ByteNone))
+          }
+        },
+        test("invalid Byte (text)") {
+          check(genAlphaLowerString)(b => assert(SafeNumbers.byte(b.toString))(equalTo(ByteNone)))
+        }
+      ),
       suite("Double")(
         test("valid") {
           check(Gen.double.filterNot(_.isNaN)) { d =>
@@ -102,8 +113,25 @@ object SafeNumbersSpec extends ZIOSpecDefault {
         test("valid (from Long)") {
           check(Gen.long)(i => assert(SafeNumbers.double(i.toString))(equalTo(DoubleSome(i.toDouble))))
         },
+        test("valid (from BigDecimal)") {
+          check(genBigDecimal)(i => assert(SafeNumbers.double(i.toString))(equalTo(DoubleSome(i.doubleValue()))))
+        },
         test("invalid edge cases") {
-          val inputs = List("N", "Inf", "-NaN", "+NaN", "e1", "1.1.1", "1 ")
+          val inputs = List(
+            "N",
+            "Inf",
+            "Info",
+            "-NaN",
+            "+NaN",
+            "e1",
+            "1.1.1",
+            "1 ",
+            "1eO",
+            "1e+2147483648",
+            "1e+3147483648",
+            "9" * 99,
+            "0." + "9" * 99
+          )
 
           check(Gen.fromIterable(inputs))(i => assert(SafeNumbers.double(i))(equalTo(DoubleNone)))
         },
@@ -116,13 +144,31 @@ object SafeNumbersSpec extends ZIOSpecDefault {
             "-0.0", // zeroes
             "0000.1",
             "0.00001",
+            "0.0e-12",
+            "1.1e-12",
+            "1.1e-1234",
+            "1.1e+1234",
             "000.00001000", // trailing zeros
             "NaN",
             "92233720368547758070", // overflows a Long significand
             "Infinity",
             "+Infinity",
             "-Infinity",
-            "3.976210887433566E-281" // rounds if a naive scaling is used
+            "503599627370496E+13",    // fast path
+            "503599627370496E+23",    // fast path with slop
+            "3.976210887433566E-281", // rounds if a naive scaling is used
+            "9007199254740993.0",     // round-down, halfway
+            "18014398509481986.0",
+            "9223372036854776832.0",
+            "9007199254740995.0", // round-up, halfway
+            "18014398509481990.0",
+            "9223372036854778880.0",
+            "9223372036854776833.0", // round-up, above halfway
+            "36028797018963967.0",   // 2^n - 1 integer regression
+            "2.2250738585072014E-308",
+            "2.2250738585072013E-308",
+            "2.2250738585072012E-308",
+            "2.2250738585072011E-308"
           )
 
           check(Gen.fromIterable(inputs)) { s =>
@@ -157,7 +203,20 @@ object SafeNumbersSpec extends ZIOSpecDefault {
           check(Gen.long)(i => assert(SafeNumbers.float(i.toString))(equalTo(FloatSome(i.toFloat))))
         },
         test("invalid edge cases") {
-          val inputs = List("N", "Inf", "-NaN", "+NaN", "e1", "1.1.1")
+          val inputs = List(
+            "N",
+            "Inf",
+            "Info",
+            "-NaN",
+            "+NaN",
+            "e1",
+            "1.1.1",
+            "1eO",
+            "1e+2147483648",
+            "1e+3147483648",
+            "9" * 99,
+            "0." + "9" * 99
+          )
 
           check(Gen.fromIterable(inputs))(i => assert(SafeNumbers.float(i))(equalTo(FloatNone)))
         },
@@ -170,12 +229,39 @@ object SafeNumbersSpec extends ZIOSpecDefault {
             "-0.0", // zeroes
             "0000.1",
             "0.00001",
+            "0.0e-12",
+            "1.1e-12",
+            "1.1e-1234",
+            "1.1e+1234",
             "000.00001000", // trailing zeros
             "NaN",
             "92233720368547758070", // overflows a Long significand
             "Infinity",
             "+Infinity",
-            "-Infinity"
+            "-Infinity",
+            "16777217.0", // round-down, halfway
+            "33554434.0",
+            "17179870208.0",
+            "16777219.0", // round-up, halfway
+            "33554438.0",
+            "17179872256.0",
+            "33554435.0", // round-up, above halfway
+            "17179870209.0",
+            "37930954282500097", // fast path with `toFloat`
+            "48696272630054913",
+            // TODO: uncomment after release of Scala Native 0.5.7
+            // "1.00000017881393432617187499", // check exactly halfway, round-up at halfway
+            // "1.000000178813934326171875",
+            // "1.00000017881393432617187501",
+            "36028797018963967.0", // 2^n - 1 integer regression
+            "1.17549435E-38",
+            "1.17549434E-38",
+            "1.17549433E-38",
+            "1.17549432E-38",
+            "1.17549431E-38",
+            "1.17549430E-38",
+            "1.17549429E-38",
+            "1.17549428E-38"
           )
 
           check(Gen.fromIterable(inputs)) { s =>
@@ -189,6 +275,9 @@ object SafeNumbersSpec extends ZIOSpecDefault {
           check(Gen.double.filterNot(_.isNaN)) { d =>
             assert(SafeNumbers.float(d.toString))(equalTo(FloatSome(d.toFloat)))
           }
+        },
+        test("valid (from BigDecimal)") {
+          check(genBigDecimal)(i => assert(SafeNumbers.float(i.toString))(equalTo(FloatSome(i.floatValue()))))
         },
         test("invalid float (text)") {
           check(genAlphaLowerString)(s => assert(SafeNumbers.float(s))(equalTo(FloatNone)))
@@ -257,7 +346,7 @@ object SafeNumbersSpec extends ZIOSpecDefault {
           check(Gen.short)(d => assert(SafeNumbers.short(d.toString))(equalTo(ShortSome(d))))
         },
         test("invalid (out of range)") {
-          check(Gen.long.filter(i => i < Short.MinValue || i > Short.MaxValue))(d =>
+          check(Gen.int.filter(i => i < Short.MinValue || i > Short.MaxValue))(d =>
             assert(SafeNumbers.short(d.toString))(equalTo(ShortNone))
           )
         },