… zio-json/js/src/main/scala/zio/json/internal/UnsafeNumbers.scala
… zio-json/jvm/src/main/scala/zio/json/internal/UnsafeNumbers.scala
… zio-json/native/src/main/scala/zio/json/internal/UnsafeNumbers.scala
… zio-json/shared/src/main/scala/zio/json/internal/numbers.scala