Tested with Scala 3 on Intel® Core™ i7-11800H:

#### Before

JDK 21:
```txt
Benchmark                          (size)   Mode  Cnt        Score       Error  Units
ArrayOfDoublesReading.zioJson         128  thrpt    5   173701.182 ±  3114.107  ops/s
ArrayOfFloatsReading.zioJson          128  thrpt    5   332838.364 ±  8347.542  ops/s
GeoJSONReading.zioJson                N/A  thrpt    5    11618.211 ±   422.198  ops/s
PrimitivesReading.zioJson             N/A  thrpt    5  3970602.729 ± 32732.282  ops/s```
```

### After

JDK 11:
```txt
Benchmark                          (size)   Mode  Cnt        Score       Error  Units
ArrayOfDoublesReading.zioJson         128  thrpt    5   279180.007 ±  1752.510  ops/s
ArrayOfFloatsReading.zioJson          128  thrpt    5   408949.463 ± 33009.599  ops/s
GeoJSONReading.zioJson                N/A  thrpt    5    11732.810 ±    64.019  ops/s
PrimitivesReading.zioJson             N/A  thrpt    5  3511617.479 ± 86705.563  ops/s
```

JDK 17:
```txt
Benchmark                          (size)   Mode  Cnt        Score       Error  Units
ArrayOfDoublesReading.zioJson         128  thrpt    5   302841.296 ±  1379.306  ops/s
ArrayOfFloatsReading.zioJson          128  thrpt    5   399378.093 ± 20565.456  ops/s
GeoJSONReading.zioJson                N/A  thrpt    5    10807.524 ±   134.091  ops/s
PrimitivesReading.zioJson             N/A  thrpt    5  3495642.795 ± 55296.299  ops/s
```

JDK 21:
```txt
Benchmark                          (size)   Mode  Cnt        Score       Error  Units
ArrayOfDoublesReading.zioJson         128  thrpt    5   245329.011 ±  2695.577  ops/s
ArrayOfFloatsReading.zioJson          128  thrpt    5   374875.316 ±  9744.821  ops/s
GeoJSONReading.zioJson                N/A  thrpt    5    11105.953 ±  1568.785  ops/s
PrimitivesReading.zioJson             N/A  thrpt    5  3937234.559 ± 65488.672  ops/s
```